Revert "rename utils to const, clean up code in Project Component"
This reverts commit 41b2ee9b3aa0ffee12ee1074eb572d98febac66a.
