+++ src/util/footer.ts
